backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.18
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
